+++ etc/org.opencastproject.statistics.export.impl.StatisticsExportServiceImpl.cfg
@@ -1,10 +1,8 @@
 # A comma separated list of mappings which map from series provider IDs to episode provider IDs. The first part of each
 # mapping has to be a series provider ID. A colon marks the beginning of the second part of the mapping which has to be
 # a corresponding episode provider ID. If these mappings are defined, then, when exporting series statistics data, the
-# data of all events of the series will be exported, instead of the series data. So you get more detailed data when
-# providing the mappings. However, if you don't provide a mapping for a certain provider, you will be able to export the
-# lower detailed series data. If you want the detailed data, Opencast needs to know which statistics provider should be
-# used on the episodes of the series you are exporting. That's why this mapping is necessary.
+# data of all events of the series will be exported, instead of the series data. If you don't provide a mapping for a
+# certain provider, you will only be able to export the lower detailed series data.
 # Example: series.to.event.provider.mappings=series.views.sum.influx:episode.views.sum.influx,foo:bar
 #
 # Default: not set
