+++ modules/admin-ui/pom.xml
@@ -75,7 +75,7 @@
     </dependency>
     <dependency>
       <groupId>org.opencastproject</groupId>
-      <artifactId>opencast-statistics-export-csv</artifactId>
+      <artifactId>opencast-statistics-export-service</artifactId>
       <version>${project.version}</version>
       <scope>provided</scope>
     </dependency>
