+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/StatisticsEndpoint.java
@@ -25,6 +25,8 @@ import static org.opencastproject.security.api.SecurityConstants.GLOBAL_ADMIN_RO
 import static org.opencastproject.util.data.functions.Functions.chuck;
 import static org.opencastproject.util.doc.rest.RestParameter.Type.STRING;
 
+import org.opencastproject.adminui.impl.ProviderQuery;
+import org.opencastproject.adminui.impl.RawProviderQuery;
 import org.opencastproject.adminui.index.AdminUISearchIndex;
 import org.opencastproject.index.service.api.IndexService;
 import org.opencastproject.index.service.impl.index.event.Event;
@@ -40,7 +42,7 @@ import org.opencastproject.statistics.api.StatisticsProvider;
 import org.opencastproject.statistics.api.StatisticsService;
 import org.opencastproject.statistics.api.TimeSeries;
 import org.opencastproject.statistics.api.TimeSeriesProvider;
-import org.opencastproject.statistics.export.api.StatisticsExportCSV;
+import org.opencastproject.statistics.export.api.StatisticsExportService;
 import org.opencastproject.util.NotFoundException;
 import org.opencastproject.util.RestUtil;
 import org.opencastproject.util.doc.rest.RestParameter;
@@ -57,11 +59,8 @@ import org.json.simple.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.time.Instant;
 import java.time.ZoneId;
-import java.time.format.DateTimeParseException;
 import java.util.Arrays;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
@@ -94,7 +93,7 @@ public class StatisticsEndpoint {
   private IndexService indexService;
   private AdminUISearchIndex searchIndex;
   private StatisticsService statisticsService;
-  private StatisticsExportCSV statisticsExportCSV;
+  private StatisticsExportService statisticsExportService;
 
   public void setSecurityService(SecurityService securityService) {
     this.securityService = securityService;
@@ -112,126 +111,8 @@ public class StatisticsEndpoint {
     this.statisticsService = statisticsService;
   }
 
-  public void setStatisticsExportCSV(StatisticsExportCSV statisticsExportCSV) {
-    this.statisticsExportCSV = statisticsExportCSV;
-  }
-
-  /**
-   * The provider query as it appears in the JSON document, to be parsed by GSON.
-   *
-   * Things like dates and enums are not validated in this class. This is done by converting it into a
-   * {@link ProviderQuery}.
-   */
-  private static final class RawProviderQuery {
-    private String providerId;
-    private String from;
-    private String to;
-    private String resourceId;
-    private String dataResolution;
-
-    // Needed for gson deserialization
-    RawProviderQuery() {
-    }
-
-    public String getProviderId() {
-      return providerId;
-    }
-
-    public String getFrom() {
-      return from;
-    }
-
-    public String getTo() {
-      return to;
-    }
-
-    public String getResourceId() {
-      return resourceId;
-    }
-
-    public String getDataResolution() {
-      return dataResolution;
-    }
-  }
-
-  private static final class ProviderQuery {
-    private String providerId;
-    private Instant from;
-    private Instant to;
-    private String resourceId;
-    private DataResolution dataResolution;
-
-    ProviderQuery(RawProviderQuery query) {
-      this.setProviderId(query.providerId);
-      this.setFrom(query.from);
-      this.setTo(query.to);
-      if (this.to.compareTo(this.from) <= 0) {
-        throw new IllegalArgumentException("'from' date must be before 'to' date");
-      }
-      this.setDataResolution(query.dataResolution);
-      this.setResourceId(query.getResourceId());
-    }
-
-    ProviderQuery(String providerId, String from, String to, String dataResolution, String resourceId) {
-      this.setProviderId(providerId);
-      this.setFrom(from);
-      this.setTo(to);
-      if (this.to.compareTo(this.from) <= 0) {
-        throw new IllegalArgumentException("'from' date must be before 'to' date");
-      }
-      this.setDataResolution(dataResolution);
-      this.setResourceId(resourceId);
-    }
-
-    private void setProviderId(String providerId) {
-      this.providerId = providerId;
-    }
-
-    private void setFrom(String from) {
-      if (isNotIso8601Utc(from)) {
-        throw new IllegalArgumentException("Missing value for 'from' or not in ISO 8601 UTC format");
-      }
-      this.from = Instant.parse(from);
-    }
-
-    private void setTo(String to) {
-      if (isNotIso8601Utc(to)) {
-        throw new IllegalArgumentException("Missing value for 'to' or not in ISO 8601 UTC format");
-      }
-      this.to = Instant.parse(to);
-    }
-
-    private void setResourceId(String resourceId) {
-      this.resourceId = resourceId;
-    }
-
-    private void setDataResolution(String dataResolution) {
-      Optional<DataResolution> resolution = dataResolutionFromJson(dataResolution);
-      if (!resolution.isPresent()) {
-        throw new IllegalArgumentException("illegal value for 'resolution'");
-      }
-      this.dataResolution = resolution.get();
-    }
-
-    public String getProviderId() {
-      return providerId;
-    }
-
-    public Instant getFrom() {
-      return from;
-    }
-
-    public Instant getTo() {
-      return to;
-    }
-
-    public String getResourceId() {
-      return resourceId;
-    }
-
-    public DataResolution getDataResolution() {
-      return dataResolution;
-    }
+  public void setStatisticsExportService(StatisticsExportService statisticsExportService) {
+    this.statisticsExportService = statisticsExportService;
   }
 
   @GET
@@ -296,23 +177,6 @@ public class StatisticsEndpoint {
     return dataResolution.toString().toLowerCase();
   }
 
-  private static Optional<DataResolution> dataResolutionFromJson(final String dataResolution) {
-    try {
-      return Optional.of(Enum.valueOf(DataResolution.class, dataResolution.toUpperCase()));
-    } catch (IllegalArgumentException e) {
-      return Optional.empty();
-    }
-  }
-
-  private static boolean isNotIso8601Utc(final String value) {
-    try {
-      Instant.parse(value);
-      return false;
-    } catch (DateTimeParseException e) {
-      return true;
-    }
-  }
-
   @POST
   @Path("data.json")
   @Produces(MediaType.APPLICATION_JSON)
@@ -342,12 +206,12 @@ public class StatisticsEndpoint {
         .map(ProviderQuery::new)
         .flatMap(q ->
           statisticsService
-            .getProvider(q.providerId)
+            .getProvider(q.getProviderId())
             .map(Stream::of).orElseGet(Stream::empty)
-            .peek(p -> checkAccess(q.resourceId, p.getResourceType()))
+            .peek(p -> checkAccess(q.getResourceId(), p.getResourceType()))
             .map(p -> timeSeriesToJson(
               p.getId(),
-              statisticsService.getTimeSeriesData(p, q.resourceId, q.from, q.to, q.dataResolution,
+              statisticsService.getTimeSeriesData(p, q.getResourceId(), q.getFrom(), q.getTo(), q.getDataResolution(),
                 ZoneId.systemDefault()))))
         .forEach(result::add);
     } catch (IllegalArgumentException e) {
@@ -379,9 +243,9 @@ public class StatisticsEndpoint {
       final ProviderQuery q = new ProviderQuery(providerId, fromStr, toStr, dataResolutionStr, resourceId);
       final StatisticsProvider p = statisticsService
         .getProvider(providerId).orElseThrow(() -> new IllegalArgumentException("Unknown provider: " + providerId));
-      checkAccess(q.resourceId, p.getResourceType());
-      final String csv = statisticsExportCSV.getCSV(p, q.resourceId, q.from, q.to, q.dataResolution, searchIndex,
-        ZoneId.systemDefault());
+      checkAccess(q.getResourceId(), p.getResourceType());
+      final String csv = statisticsExportService.getCSV(p, q.getResourceId(), q.getFrom(), q.getTo(), q.getDataResolution(),
+        searchIndex, ZoneId.systemDefault());
       return Response.ok().entity(csv).build();
     } catch (IllegalArgumentException e) {
       return RestUtil.R.badRequest(e.getMessage());
